+++ routes/api.php
@@ -23,5 +23,8 @@ Route::middleware('auth:sanctum')->group(function () {
     });
     Route::apiResource('sinks', SinkApiController::class)->only(['index', 'show', 'update']);
     Route::apiResource('sinks.chunks', ChunkController::class)->only(['index', 'update']);
+    Route::controller(ChunkController::class)->group(function () {
+        Route::get('sinks/{sink}/chunks/{chunk}/download', 'download');
+    });
     Route::apiResource('batch', BatchApiController::class)->only(['index', 'show', 'destroy']);
 });